services/pyproject.toml
[tool.poetry]
name = "aithena-services"
version = "0.1.0-dev1"
description = "A tool for integrating and organizing interactions with LLMs and embedding models."
authors = ["Camilo Velez <camilo.velezramirez@nih.gov>, Antoine Gerardin <antoine.gerardin@nih.gov>"]
readme = "README.md"
packages = [{include = "polus", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pydantic = "^2.7.4"
openai = "^1.35.13"
requests = "^2.32.3"
ollama = "^0.3.1"
fastapi = {extras = ["standard"], version = ">=0.112.0"}
pytest-asyncio = "^0.23.8"
python-dotenv = "^1.0.1"
llama-index = "^0.11.0"
llama-index-embeddings-azure-openai = "^0.2.1"
llama-index-embeddings-ollama = "^0.3.0"
llama-index-llms-openai = "^0.2.0"
llama-index-llms-ollama = "^0.3.0"
llama-index-llms-azure-openai = "^0.2.0"
typing-extensions = "^4.12.2"
solara = "^1.39.0"
uvicorn = "^0.30.6"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
#log_file = "pytest.log"
#log_file_level = "INFO"
#log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
#log_file_date_format = "%Y-%m-%d %H:%M:%S"
